added navigation from sectionlist to different pages
