+++ App.js
@@ -1,89 +1,21 @@
 import React, { Component } from 'react';
-import { StyleSheet, Text, View } from 'react-native';
-import Accordion from 'react-native-collapsible/Accordion';
+import { NavigationContainer } from '@react-navigation/native';
+import { createStackNavigator } from '@react-navigation/stack';
 
-export default class MyComponent extends React.Component {
-  constructor(props) {
-    super(props);
+import SectionList from "./components/SectionList";
+import Details from "./components/Details";
 
-    this.state = {
-      sections: sections,
-      activeSections: [],
-    };
+const Stack = createStackNavigator();
 
-    this.renderHeader = this.renderHeader.bind(this);
-    this.renderContent = this.renderContent.bind(this);
-    this.updateSections = this.updateSections.bind(this);
-  }
-
-  updateSections(activeSections) {
-    this.setState({ activeSections });
-  }
-
-  renderHeader(section) {
-    return (
-      <View style={styles.header}>
-        <Text style={styles.title}>{section.title}</Text>
-      </View>
-    );
-  }
-
-  renderContent(section) {
-    return (
-      <View>
-        <Text>{section.content}</Text>
-      </View>
-    );
-  }
-
+export default class MyComponent extends Component {
   render() {
     return (
-      <View>
-        <Accordion
-          sections={this.state.sections}
-          activeSections={this.state.activeSections}
-          renderHeader={this.renderHeader}
-          renderContent={this.renderContent}
-          onChange={this.updateSections}
-         />
-      </View>
+      <NavigationContainer>
+        <Stack.Navigator>
+          <Stack.Screen name="Section List" component={SectionList} />
+          <Stack.Screen name="Details" component={Details} />
+        </Stack.Navigator>
+      </NavigationContainer>
     );
   }
 }
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: '#F5FCFF',
-  },
-  title: {
-    textAlign: 'center',
-    fontSize: 22,
-    fontWeight: '300',
-    marginBottom: 20,
-  },
-  header: {
-    backgroundColor: '#F5FCFF',
-    padding: 10,
-  },
-  headerText: {
-    textAlign: 'center',
-    fontSize: 16,
-    fontWeight: '500',
-  },
-  content: {
-    padding: 20,
-    backgroundColor: '#fff',
-  },
-});
-
-const sections = [
-  {
-    title: 'First',
-    content: 'Lorem ipsum...',
-  },
-  {
-    title: 'Second',
-    content: 'Lorem ipsum...',
-  },
-];
