+++ components/Details.js
@@ -1,0 +1,16 @@
+import React, { Component } from "react";
+import { View, Text } from "react-native";
+
+export default class DetailTwo extends Component {
+  constructor(props) {
+    super(props);
+  }
+
+  render() {
+    return (
+      <View style={{ flex: 1, alignItems: "center", justifyContent: "center" }}>
+        <Text>Details for clicking an item in sectionlist. { this.props.route.params.whichDetail }</Text>
+      </View>
+    );
+  }
+}
